Chat-App/chat_api/requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.2
django-cors-headers==4.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
